Rework registry to combine provider and registry interface, move capabilities out and enable Registry injection in Builder
